+++ packages/typespec-client-generator-core/src/internal-utils.ts
@@ -1,4 +1,9 @@
-import { isPreviewVersion } from "@azure-tools/typespec-azure-core";
+import {
+  FinalStateValue,
+  getLroMetadata,
+  isPreviewVersion,
+  LroMetadata,
+} from "@azure-tools/typespec-azure-core";
 import {
   BooleanLiteral,
   compilerAssert,
@@ -43,6 +48,7 @@ import {
   getAlternateType,
   getClientDocExplicit,
   getClientLocation,
+  getMarkAsLro,
   getParamAlias,
 } from "./decorators.js";
 import { getSdkHttpParameter, isSdkHttpParameter } from "./http.js";
@@ -51,15 +57,18 @@ import {
   ExternalTypeInfo,
   SdkBuiltInType,
   SdkClient,
+  SdkClientType,
   SdkEnumType,
   SdkHeaderParameter,
   SdkHttpResponse,
   SdkMethodParameter,
   SdkOperationGroup,
+  SdkServiceOperation,
   SdkType,
   TCGCContext,
 } from "./interfaces.js";
 import { createDiagnostic, createStateSymbol } from "./lib.js";
+import { getSdkBasicServiceMethod } from "./methods.js";
 import {
   getCrossLanguageDefinitionId,
   getDefaultApiVersion,
@@ -872,4 +881,48 @@ export function findEntriesWithTarget<TSource extends Type, TTarget>(
     }
   }
   return results;
+}
+
+/**
+ * Retrieves Long Running Operation (LRO) metadata for a given operation.
+ *
+ * This function serves as a wrapper that:
+ * 1. First tries to get LRO metadata using the `getLroMetadata` function from the Azure Core library
+ * 2. If unavailable or undefined, it would check for the existence of a `getMarkAsLro` function
+ *    and return a mock LRO metadata object if the operation is marked as LRO
+ *
+ * @param context - The TypeSpec client generator context
+ * @param operation - The TypeSpec operation to check for LRO metadata
+ * @returns The LRO metadata for the operation if available, otherwise undefined
+ */
+export function getTcgcLroMetadata<TServiceOperation extends SdkServiceOperation>(
+  context: TCGCContext,
+  operation: Operation,
+  client: SdkClientType<TServiceOperation>,
+): LroMetadata | undefined {
+  const lroMetaData = getLroMetadata(context.program, operation);
+  if (lroMetaData) {
+    return lroMetaData;
+  }
+  if (getMarkAsLro(context, operation)) {
+    // we guard against this in the setting of `@markAsLro`
+    const sdkMethod = ignoreDiagnostics(getSdkBasicServiceMethod(context, operation, client));
+    const returnType = sdkMethod.response.type!.__raw! as Model;
+    return {
+      operation,
+      logicalResult: returnType,
+      finalStateVia: FinalStateValue.originalUri, // doesn't really matter, but for get LRO
+      pollingInfo: {
+        kind: "pollingOperationStep",
+        responseModel: returnType,
+        terminationStatus: {
+          kind: "status-code",
+        },
+      },
+      envelopeResult: returnType,
+      finalEnvelopeResult: returnType,
+      finalResult: returnType,
+    };
+  }
+  return undefined;
 }